Connect backend to frontend
Adding, Modifying, Searching and Deleting lines now functions in the app,
along with the initial Configuration of the app. The Config table has a default value
that can be updated by the user.

These functions have also been renamed in the backend code for better
readability.

Moved and added some TODOs.

GH-4