Update categories, account tables
Changed categories table select in backend to only use the category name
and leave the rest of the columns out of the select.

Also added a query that can get the category id by category name.

Account table now has NOT NULL constraints and the amount
column is decimal to store floating point numbers.

GH-4